Side benefit of REST rework: we can have smaller files now. Break up endpoint specific stuff from user-data into endpoint.js
