+++ src-electron/rest/user-data.js
@@ -45,44 +45,6 @@ function httpGetSessionKeyValues(db) {
 }
 
 /**
- * HTTP DELETE: endpoint
- *
- * @param {*} db
- * @returns callback for the express uri registration
- */
-function httpDeleteEndpoint(db) {
-  return (request, response) => {
-    var id = request.query.id
-    queryConfig.deleteEndpoint(db, id).then((removed) => {
-      response.json({
-        successful: removed > 0,
-        id: id,
-      })
-      return response.status(restApi.httpCode.ok).send()
-    })
-  }
-}
-
-/**
- * HTTP DELETE: endpoint type
- *
- * @param {*} db
- * @returns callback for the express uri registration
- */
-function httpDeleteEndpointType(db) {
-  return (request, response) => {
-    var id = request.query.id
-    queryConfig.deleteEndpointType(db, id).then((removed) => {
-      response.json({
-        successful: removed > 0,
-        id: id,
-      })
-      return response.status(restApi.httpCode.ok).send()
-    })
-  }
-}
-
-/**
  * HTTP POST: save session key value
  *
  * @param {*} db
@@ -262,66 +224,6 @@ function httpPostCommandUpdate(db) {
 }
 
 /**
- * HTTP POST: endpoint
- *
- * @param {*} db
- * @returns callback for the express uri registration
- */
-function httpPostEndpoint(db) {
-  return (request, response) => {
-    var { endpointId, networkId, endpointType } = request.body
-    var sessionIdexport = request.session.zapSessionId
-    queryConfig
-      .insertEndpoint(db, sessionIdexport, endpointId, endpointType, networkId)
-      .then((newId) =>
-        validation.validateEndpoint(db, newId).then((validationData) => {
-          response.json({
-            id: newId,
-            endpointId: endpointId,
-            endpointType: endpointType,
-            networkId: networkId,
-            validationIssues: validationData,
-          })
-        })
-      )
-      .catch((err) => {
-        return response.status(restApi.httpCode.badRequest).send()
-      })
-  }
-}
-
-/**
- * HTTP POST: endpoint
- *
- * @param {*} db
- * @returns callback for the express uri registration
- */
-function httpPatchEndpoint(db) {
-  return (request, response) => {
-    var { context } = request.body
-    var sessionIdexport = request.session.zapSessionId
-    let changes = context.changes.map((data) => {
-      var paramType = ''
-      return { key: data.updatedKey, value: data.value, type: paramType }
-    })
-
-    queryConfig
-      .updateEndpoint(db, sessionIdexport, context.id, changes)
-      .then((data) => {
-        return validation
-          .validateEndpoint(db, context.id)
-          .then((validationData) => {
-            response.json({
-              endpointId: context.id,
-              changes: context.changes,
-              validationIssues: validationData,
-            })
-            return response.status(restApi.httpCode.ok).send()
-          })
-      })
-  }
-}
-/**
  * HTTP GET: initial state
  *
  * @param {*} db
@@ -355,59 +257,6 @@ function httpGetInitialState(db) {
     Promise.all(statePopulators).then(() => {
       return response.status(restApi.httpCode.ok).json(state)
     })
-  }
-}
-
-/**
- * HTTP POST endpoint type
- *
- * @param {*} db
- * @returns callback for the express uri registration
- */
-function httpPostEndpointType(db) {
-  return (request, response) => {
-    var { name, deviceTypeRef } = request.body
-    var sessionId = request.session.zapSessionId
-    queryConfig
-      .insertEndpointType(db, sessionId, name, deviceTypeRef)
-      .then((newId) =>
-        response.json({
-          id: newId,
-          name: name,
-          deviceTypeRef: deviceTypeRef,
-        })
-      )
-      .catch((err) => response.status(restApi.httpCode.badRequest).send())
-  }
-}
-
-/**
- * HTTP POST: endpoint type update
- *
- * @param {*} db
- * @returns callback for the express uri registration
- */
-function httpPatchEndpointType(db) {
-  return (request, response) => {
-    var { endpointTypeId, updatedKey, updatedValue } = request.body
-    var sessionId = request.session.zapSessionId
-
-    queryConfig
-      .updateEndpointType(
-        db,
-        sessionId,
-        endpointTypeId,
-        updatedKey,
-        updatedValue
-      )
-      .then(() => {
-        response.json({
-          endpointTypeId: endpointTypeId,
-          updatedKey: updatedKey,
-          updatedValue: updatedValue,
-        })
-        return response.status(restApi.httpCode.ok).send()
-      })
   }
 }
 
@@ -451,25 +300,6 @@ exports.post = [
     uri: restApi.uri.saveSessionKeyValue,
     callback: httpPostSaveSessionKeyValue,
   },
-  {
-    uri: restApi.uri.endpoint,
-    callback: httpPostEndpoint,
-  },
-  {
-    uri: restApi.uri.endpointType,
-    callback: httpPostEndpointType,
-  },
-]
-
-exports.patch = [
-  {
-    uri: restApi.uri.endpoint,
-    callback: httpPatchEndpoint,
-  },
-  {
-    uri: restApi.uri.endpointType,
-    callback: httpPatchEndpointType,
-  },
 ]
 
 exports.get = [
@@ -484,16 +314,5 @@ exports.get = [
   {
     uri: `${restApi.uri.option}/:category`,
     callback: httpGetOption,
-  },
-]
-
-exports.delete = [
-  {
-    uri: restApi.uri.endpoint,
-    callback: httpDeleteEndpoint,
-  },
-  {
-    uri: restApi.uri.endpointType,
-    callback: httpDeleteEndpointType,
   },
 ]
